Minor cleanup

+++ log4j2-elasticsearch-core/src/test/java/org/appenders/log4j2/elasticsearch/failover/SingleKeySequenceSelectorTest.java
@@ -316,8 +316,8 @@ public class SingleKeySequenceSelectorTest {
 
     }
 
-    private void sleep(int millis) {
-        LockSupport.parkNanos(millis * 1000000);
+    private void sleep(long millis) {
+        LockSupport.parkNanos(millis * 1000000L);
     }
 
 }
